requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
google-generativeai==0.8.3
requests>=2.32.1
beautifulsoup4==4.12.0
pdfplumber==0.11.0
python-multipart==0.0.9
python-dotenv==1.0.0
